+++ src/modules/documents/documents.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 import { DocumentsController } from './documents.controller';
 import { MinioService } from '../../common/aws/minio.service';
 import { DocumentsRepository } from './repository/document.repository';
-import { CreateDocumentService } from './services/create_document.service';
-import { FindAllDocumentsService } from './services/find_all_documents.service';
+import { CreateDocumentService } from './services/create-document.service';
+import { FindAllDocumentsService } from './services/find-all-documents.service';
 
 @Module({
   controllers: [DocumentsController],
